follow ide advices

+++ src/main/java/vavi/swing/border/BevelBorderCustomizer.java
@@ -42,7 +42,7 @@ public class BevelBorderCustomizer extends BasicBorderCustomizer {
 
             String name = ev.getPropertyName();
             if ("bevelType".equals(name)) {
-                type = ((Integer) ev.getNewValue()).intValue();
+                type = (Integer) ev.getNewValue();
             }
             else if ("highlightInnerColor".equals(name)) {
                 hi = (Color) ev.getNewValue();
